+++ src/lib/models/reflections/index.ts
@@ -1,14 +1,14 @@
 export {
     Reflection,
-    ReflectionKind,
     ReflectionFlag,
-    TraverseProperty,
     ReflectionFlags,
+    TraverseProperty,
 } from "./abstract";
 export type { Decorator, TraverseCallback } from "./abstract";
 export { ContainerReflection } from "./container";
 export { DeclarationReflection } from "./declaration";
 export type { DeclarationHierarchy } from "./declaration";
+export { ReflectionKind } from "./kind";
 export { ParameterReflection } from "./parameter";
 export { ProjectReflection } from "./project";
 export { ReferenceReflection } from "./reference";
